Add files via upload
Add CTFFIND5 form. Change CTFFIND4 form for optional run of ace. Defaults set for CTFFIND5 are what we found work best at NYU. I configured the branch to NOT require any changes to the appion database, since tilt and tilt axis already existed. Sample thickness can be determined but will not be added to the database as currently configured. If it is desired, there are lines noted in the code which can be uncommented. In our experience at NYU, the program often gives slightly higher resolution determination with thickness turned on, presumably because of the extra parameter to fit. However, it often does not match the thickness determined by ALS or ZLP. Additionally, it can sometimes find negative values or unrealistically high values. Although we are saving to the database, we generally do not find it helpful.